win: fix sqInt LLP64 truncation (was typedef long) — root cause of corruption
sqInt/usqInt/sqLong/sqIntptr_t were typedef'd to long, which is 64-bit on LP64
(Linux/macOS/ARM64) but only 32-bit on Windows (LLP64). sqInt holds oops and
pointers throughout the plugin/InterpreterProxy interface, so on Windows every
oop above 4 GB (the heap is at ~2.7 TiB) got truncated to 32 bits and
sign-extended into a garbage pointer -> heap corruption -> crash. Pinned via a
new Windows crash-dump Vectored Exception Handler (test_load_image.cpp): the
fault address was a sign-extended 32-bit value and addr2line put the fault in
proxy_isBytes(long)/ObjectHeader::isBytesObject.

Fix: typedef sqInt/etc. to intptr_t/int64_t (pointer-sized on all 64-bit
hosts; byte-identical to long on LP64, so Linux/macOS unchanged). Verified:
CoCompletionEngineTest 65/65 (was a hard SIGSEGV), 996/996 sanity across
SmallInteger/Array/Dictionary/String (no regression). The Windows crash-dump
VEH (fault addr + rip + rva + native backtrace; CONTINUE_SEARCH, no recovery)
replaces the no-diagnostics default on Windows.

+++ src/vm/plugins/sqMemoryAccess.h
@@ -9,14 +9,20 @@
 #define SQ_IMAGE64 1
 #define SQ_HOST64 1
 
-typedef long sqInt;
-typedef unsigned long usqInt;
+/* sqInt MUST be pointer-sized (it holds oops/pointers).  `long` is 64-bit on
+ * LP64 (Linux/macOS/ARM64) but only 32-bit on Windows (LLP64), which truncated
+ * every oop passed through the plugin/InterpreterProxy interface (e.g.
+ * proxy_isBytes(sqInt) -> garbage ObjectHeader* -> crash).  intptr_t/int64_t
+ * are 64-bit on all 64-bit hosts and identical to `long` on LP64, so this is
+ * byte-for-byte unchanged on Linux/macOS and fixes Windows. */
+typedef intptr_t sqInt;
+typedef uintptr_t usqInt;
 
-#define sqLong long
-#define usqLong unsigned long
+#define sqLong int64_t
+#define usqLong uint64_t
 
-typedef long sqIntptr_t;
-typedef unsigned long usqIntptr_t;
+typedef intptr_t sqIntptr_t;
+typedef uintptr_t usqIntptr_t;
 
 /* Memory access macros - direct access since same endianness */
 #define byteAtPointer(ptr)       ((unsigned char)(*(ptr)))
